refactor(desktop): Consolidate platform-specific components for desktop and electron
- Rename `.desktop.tsx` files to `.electron.tsx` for consistent naming
- Update import paths in main component files to use `.electron.tsx`
- Remove redundant files and consolidate platform-specific implementations
- Adjust import resolver for platform-specific components
