+++ apps/desktop/src/renderer/src/modules/entry-column/layouts/EntryListHeader.tsx
@@ -1,6 +1,6 @@
 import { withResponsiveSyncComponent } from "@follow/components/utils/selector.js"
 
-import { EntryListHeader as EntryListHeaderDesktop } from "./EntryListHeader.desktop"
+import { EntryListHeader as EntryListHeaderDesktop } from "./EntryListHeader.electron"
 import { EntryListHeader as EntryListHeaderMobile } from "./EntryListHeader.mobile"
 import type { EntryListHeaderProps } from "./EntryListHeader.shared"
