+++ apps/desktop/src/renderer/src/modules/entry-column/wrapper.tsx
@@ -1,6 +1,6 @@
 import { withResponsiveSyncComponent } from "@follow/components/utils/selector.js"
 
-import { EntryColumnWrapper as EntryColumnWrapperDesktop } from "./wrapper.desktop"
+import { EntryColumnWrapper as EntryColumnWrapperDesktop } from "./wrapper.electron"
 import { EntryColumnWrapper as EntryColumnWrapperMobile } from "./wrapper.mobile"
 import type { EntryColumnWrapperProps } from "./wrapper.shared"
