+++ apps/desktop/src/renderer/src/modules/entry-content/components/EntryReadHistory.tsx
@@ -1,6 +1,6 @@
 import { withResponsiveSyncComponent } from "@follow/components/utils/selector.js"
 
-import { EntryReadHistory as EntryReadHistoryDesktop } from "./EntryReadHistory.desktop"
+import { EntryReadHistory as EntryReadHistoryDesktop } from "./EntryReadHistory.electron"
 import { EntryReadHistory as EntryReadHistoryMobile } from "./EntryReadHistory.mobile"
 
 export const EntryReadHistory = withResponsiveSyncComponent(
