+++ apps/desktop/src/renderer/src/modules/entry-content/header.tsx
@@ -1,6 +1,6 @@
 import { withResponsiveSyncComponent } from "@follow/components/utils/selector.js"
 
-import { EntryHeader as EntryHeaderDesktop } from "./header.desktop"
+import { EntryHeader as EntryHeaderDesktop } from "./header.electron"
 import { EntryHeader as EntryHeaderMobile } from "./header.mobile"
 import type { EntryHeaderProps } from "./header.shared"
