+++ apps/desktop/src/renderer/src/modules/power/transaction-section/tx-table.tsx
@@ -1,6 +1,6 @@
 import { withResponsiveSyncComponent } from "@follow/components/utils/selector.js"
 
-import { TxTable as TxTableDesktop } from "./tx-table.desktop"
+import { TxTable as TxTableDesktop } from "./tx-table.electron"
 import { TxTable as TxTableMobile } from "./tx-table.mobile"
 import type { TxTableProps } from "./tx-table.shared"
