+++ apps/desktop/src/renderer/src/modules/profile/user-profile-modal.tsx
@@ -1,6 +1,6 @@
 import { withResponsiveSyncComponent } from "@follow/components/utils/selector.js"
 
-import { UserProfileModalContent as UserProfileModalContentDesktop } from "./user-profile-modal.desktop"
+import { UserProfileModalContent as UserProfileModalContentDesktop } from "./user-profile-modal.electron"
 import { UserProfileModalContent as UserProfileModalContentMobile } from "./user-profile-modal.mobile"
 import type { SubscriptionModalContentProps } from "./user-profile-modal.shared"
