+++ apps/desktop/src/renderer/src/modules/timeline-column/FeedList.tsx
@@ -1,6 +1,6 @@
 import { withResponsiveSyncComponent } from "@follow/components/utils/selector.js"
 
-import { FeedList as FeedListDesktop } from "./FeedList.desktop"
+import { FeedList as FeedListDesktop } from "./FeedList.electron"
 import { FeedList as FeedListMobile } from "./FeedList.mobile"
 
 export const FeedList = withResponsiveSyncComponent(FeedListDesktop, FeedListMobile)
